refactor(collectors): Phase 3.2 — extract collector/workers to backend/collectors/
Pure move: host_poller, uptime_worker, sample_once, _rollup_now, _rollup_net_now,
collector, brief_worker moved to backend/collectors/__init__.py with lazy
import app as _app for all globals. app.py re-exports all moved names.
Snapshot suite confirms API responses unchanged.
